update dtype for Bert

+++ lambda-optimize/arm_tvm/lambda_function.py
@@ -57,7 +57,7 @@ def load_model(framework,model_name,model_size):
    
     return model
 
-def optimize_tvm(wtype,framework,model_name,batchsize,model_size,imgsize=224,seq_length=128, layout="NCHW",dtype="float32"):
+def optimize_tvm(wtype,framework,model_name,batchsize,model_size,imgsize=224,seq_length=128, layout="NCHW"):
     import tvm
     from tvm import relay
 
@@ -77,8 +77,8 @@ def optimize_tvm(wtype,framework,model_name,batchsize,model_size,imgsize=224,seq
 
     # NLP input 
     elif wtype == "nlp":
-        inputs = np.random.randint(0, 2000, size=(batchsize, seq_length)).astype(dtype)
-        token_types = np.random.uniform(size=(batchsize, seq_length)).astype(dtype)
+        inputs = np.random.randint(0, 2000, size=(batchsize, seq_length)).astype("int")
+        token_types = np.random.uniform(size=(batchsize, seq_length)).astype("int")
 
         tokens_tensor = torch.tensor(np.array(inputs))
         segments_tensors = torch.tensor(np.array(token_types))
